Convert /update_lang/ to use Bootstrap
LEARNER-1922
