+++ cms/static/sass/elements/_system-feedback.scss
@@ -72,6 +72,60 @@
   }
 }
 
+.page-banner {
+    max-width: $fg-max-width;
+    margin: 0 auto;
+
+    .user-messages {
+        padding-top: $baseline;
+
+        // Hack: force override the global important rule
+        // that courseware links don't have an underline.
+        a:hover {
+            color: $link-color;
+            text-decoration: underline !important;
+        }
+    }
+
+    .alert {
+        margin-bottom: $baseline !important;
+        padding: $baseline;
+        border: 1px solid;
+
+        .icon-alert {
+            margin-right: $baseline / 4;
+        }
+
+        &.alert-info {
+            color: $state-info-text;
+            background-color: $state-info-bg;
+            border-color: $state-info-border;
+            box-shadow: none;
+        }
+
+        &.alert-success {
+            color: $state-success-text;
+            background-color: $state-success-bg;
+            border-color: $state-success-border;
+            box-shadow: none;
+        }
+
+        &.alert-warning {
+            color: $state-warning-text;
+            background-color: $state-warning-bg;
+            border-color: $state-warning-border;
+            box-shadow: none;
+        }
+
+        &.alert-danger {
+            color: $state-danger-text;
+            background-color: $state-danger-bg;
+            border-color: $state-danger-border;
+            box-shadow: none;
+        }
+    }
+}
+
 .alert, .notification, .prompt {
 
   // types - confirm
