+++ cms/static/sass/partials/_variables.scss
@@ -241,6 +241,7 @@ $ui-action-primary-color-focus: $blue-s1 !default;
 
 $ui-link-color: $blue-u2 !default;
 $ui-link-color-focus: $blue-s1 !default;
+$link-color: $ui-link-color;
 
 // +Specific UI
 // ====================
@@ -281,3 +282,24 @@ $action-primary-active-bg: #1AA1DE !default; // $m-blue
 $very-light-text: $white !default;
 
 $color-background-alternate: rgb(242, 248, 251) !default;
+
+
+// ----------------------------
+// #COLORS- Bootstrap-style
+// ----------------------------
+
+$state-success-text: $black !default;
+$state-success-bg: #dff0d8 !default;
+$state-success-border: darken($state-success-bg, 5%) !default;
+
+$state-info-text: $black !default;
+$state-info-bg: #d9edf7 !default;
+$state-info-border: darken($state-info-bg, 7%) !default;
+
+$state-warning-text: $black !default;
+$state-warning-bg: #fcf8e3 !default;
+$state-warning-border: darken($state-warning-bg, 5%) !default;
+
+$state-danger-text: $black !default;
+$state-danger-bg: #f2dede !default;
+$state-danger-border: darken($state-danger-bg, 5%) !default;
