+++ cms/templates/base.html
@@ -9,9 +9,11 @@
 <%!
 from django.utils.translation import ugettext as _
 
+from openedx.core.djangoapps.util.user_messages import user_messages
 from openedx.core.djangolib.js_utils import (
     dump_js_escaped_json, js_escaped_string
 )
+from openedx.core.djangolib.markup import HTML
 %>
 
 <%page expression_filter="h"/>
@@ -69,15 +71,32 @@ from openedx.core.djangolib.js_utils import (
 
     <!-- view -->
     <div class="wrapper wrapper-view" dir="${static.dir_rtl()}">
-        <% online_help_token = self.online_help_token() if hasattr(self, 'online_help_token') else None %>
-        <%include file="widgets/header.html" args="online_help_token=online_help_token" />
+      <% online_help_token = self.online_help_token() if hasattr(self, 'online_help_token') else None %>
+      <%include file="widgets/header.html" args="online_help_token=online_help_token" />
+
+      <%
+      banner_messages = list(user_messages(request))
+      %>
+
+      % if banner_messages:
+        <div class="page-banner">
+          <div class="user-messages">
+            % for message in banner_messages:
+              <div class="alert ${message.css_class}" role="alert">
+                <span class="icon icon-alert fa ${message.icon_class}" aria-hidden="true"></span>
+                ${HTML(message.message_html)}
+              </div>
+            % endfor
+          </div>
+        </div>
+      % endif
 
       <div id="page-alert">
       <%block name="page_alert"></%block>
       </div>
 
       <main id="main" aria-label="Content" tabindex="-1">
-          <div id="content">
+          <div id="content" class="content">
           <%block name="content"></%block>
           </div>
       </main>
