+++ cms/templates/darklang/preview_lang.html
@@ -1,5 +1,3 @@
-## Override the default styles_version to the Pattern Library version (version 2)
-<%! main_css = "style-main-v2" %>
 <%page expression_filter="h"/>
 <%!
 from django.utils.translation import ugettext as _
