+++ cms/templates/ux/reference/index.html
@@ -18,18 +18,20 @@
             <article class="window unit-body">
                 <h1>UX Style Reference</h1>
 
-                <ol class="components">
-                    <li class="component">
-                        <div class="wrapper wrapper-component-action-header">
-                            <h2>Page Types</h2>
-                        </div>
-                        <section class="xblock xblock-student_view xmodule_display xmodule_HtmlModule">
-                            <ul>
-                                <li><a href="pattern-library-test.html">Pattern Library test page</a></li>
-                            </ul>
-                        </section>
-                    </li>
-                </ol>
+                <section>
+                    <h3>v1-style LMS Pages</h3>
+                    <ul>
+                    </ul>
+
+                    <h3>Pattern Library</h3>
+                    <ul>
+                        <li><a href="pattern-library/test.html">Test page</a></li>
+                    </ul>
+
+                    <h3>Bootstrap</h3>
+                    <ul>
+                    </ul>
+                </section>
             </article>
         </div>
     </div>
