+++ cms/templates/ux/reference/pattern-library/test.html
@@ -1,7 +1,7 @@
 ## Override the default styles_version to the Pattern Library version (version 2)
 <%! main_css = "style-main-v2" %>
 
-<%inherit file="../../base.html" />
+<%inherit file="/base.html" />
 
 <%block name="title">Pattern Library Test Page</%block>
 <%block name="bodyclass">is-signedin pattern-library</%block>
