+++ lms/static/sass/bootstrap/_navigation.scss
@@ -1,6 +1,6 @@
 // Local overrides for bootstrap navigation bar theming
 .navigation-container {
-  border-bottom: 1px solid $brand-primary;
+  border-bottom: 2px solid $brand-primary;
   text-decoration: none;
   background-color: $header-bg;
