+++ lms/templates/darklang/preview_lang.html
@@ -1,5 +1,8 @@
-## Override the default styles_version to the Pattern Library version (version 2)
-<%! main_css = "style-main-v2" %>
+## mako
+
+## Override the default styles_version to use Bootstrap
+<%! main_css = "css/bootstrap/lms-main.css" %>
+
 <%page expression_filter="h"/>
 <%!
 from django.utils.translation import ugettext as _
@@ -8,7 +11,7 @@ from django.utils.translation import ugettext as _
 <%inherit file="/main.html" />
 <%block name="pagetitle">${_("Preview Language Setting")}</%block>
 <%block name="nav_skip"></%block>
-<%block name="bodyclass">pattern-library</%block>
+<%block name="bodyclass">preview-language</%block>
 
 <%block name="content">
     <%include file="/dark_lang/preview_lang_include.html" />
