+++ lms/templates/ux/reference/index.html
@@ -16,7 +16,7 @@
             <article class="window unit-body">
                 <h2>UX Style Reference</h2>
 
-                <section class="xblock xblock-student_view xmodule_display xmodule_HtmlModule">
+                <section>
                     <h3>v1-style LMS Pages</h3>
                     <ul>
                         <li><a href="v1/course-skeleton.html">Course skeleton page</a></li>
