+++ openedx/core/djangoapps/dark_lang/templates/dark_lang/preview_lang_include.html
@@ -3,10 +3,10 @@ from django.utils.translation import ugettext as _
 %>
 <%page expression_filter="h"/>
 
-<h1>
+<h3>
     ${_("Preview Language Setting")}
-</h1>
-<div >
+</h3>
+<div>
     <form class="form" action="/update_lang/" method="post">
         <fieldset class="form-group">
             <div class="field">
@@ -23,14 +23,4 @@ from django.utils.translation import ugettext as _
 
         <input type="hidden" name="csrfmiddlewaretoken" value="${ csrf_token }"/>
     </form>
-
-    <br/>
-    % if not form_submit_message is UNDEFINED:
-        <h3 class="alert-title"> ${form_submit_message}</h3>
-        % if success:
-            <p class="alert-copy-with-title">
-                ${_("Please refresh the page to see the changes applied.")}
-            </p>
-        % endif
-    % endif
 </div>
